requirements.txt
google-genai>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0       # optional: only needed for --deskew
